Created mediaService file for artwork CDN access and updated file paths
