+++ Application/client/src/components/WhatsNewCarousel.js
@@ -3,8 +3,7 @@ import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Carousel from 'react-bootstrap/Carousel';
-
-import { generateImageSrc } from '../services/memberService';
+import { generateImageSrc } from '../services/mediaService';
 
 function WhatsNewCarousel({ media }) {
   const dvds = media.filter(item => item.Type === 'DVD');
